refactor: move `HasHighlightOnFocus` impl to `DwcField`
